pack buttons in w3c standard order

+++ GamepadConduitApp/SDController/mac_controller.h
@@ -36,7 +36,7 @@ struct ControllerState {
 };
 
 
-uint16_t packedButtonsState(struct ControllerState* const cs);
+uint32_t packedButtonsState(struct ControllerState* const cs);
 
 
 @protocol ControllerDriver <NSObject>
